+++ internal/examples/readme_validation_plan_example_test.go
@@ -69,7 +69,7 @@ func Example_validationPlan() {
 	).
 		WithName("Teacher")
 
-	plan := govy.Plan[Teacher](teacherValidator)
+	plan := govy.Plan(teacherValidator)
 	enc := json.NewEncoder(os.Stdout)
 	enc.SetIndent("", "  ")
 	_ = enc.Encode(plan)
