+++ pkg/govy/plan_test.go
@@ -5,7 +5,6 @@ import (
 	_ "embed"
 	"encoding/json"
 	"errors"
-	"strings"
 	"testing"
 
 	"github.com/nobl9/govy/internal/assert"
@@ -134,34 +133,13 @@ func TestPlan(t *testing.T) {
 	).
 		WithName("Pod")
 
-	t.Run("plan for Validator", func(t *testing.T) {
-		plan := govy.Plan[Pod](validator)
+	plan := govy.Plan(validator)
 
-		buf := bytes.Buffer{}
-		enc := json.NewEncoder(&buf)
-		enc.SetIndent("", "  ")
-		err := enc.Encode(plan)
-		assert.Require(t, assert.NoError(t, err))
+	buf := bytes.Buffer{}
+	enc := json.NewEncoder(&buf)
+	enc.SetIndent("", "  ")
+	err := enc.Encode(plan)
+	assert.Require(t, assert.NoError(t, err))
 
-		assert.Equal(t, expectedPlanJSON, buf.String())
-	})
-
-	t.Run("plan for ValidatorForSlice", func(t *testing.T) {
-		validatorForSlice := govy.NewForSlice(validator)
-
-		plan := govy.Plan[Pod](validatorForSlice)
-
-		buf := bytes.Buffer{}
-		enc := json.NewEncoder(&buf)
-		enc.SetIndent("", "  ")
-		err := enc.Encode(plan)
-		assert.Require(t, assert.NoError(t, err))
-
-		expectedSlicePlanJSON := expectedPlanJSON
-		i := strings.Index(expectedSlicePlanJSON, `  "properties": [`)
-		expectedSlicePlanJSON = expectedSlicePlanJSON[:i] + `  "isSlice": true,` + "\n" + expectedSlicePlanJSON[i:]
-		expectedSlicePlanJSON = strings.ReplaceAll(expectedSlicePlanJSON, `"path": "$`, `"path": "$[*]`)
-
-		assert.Equal(t, expectedSlicePlanJSON, buf.String())
-	})
+	assert.Equal(t, expectedPlanJSON, buf.String())
 }
