+++ pkg/govy/validator_test.go
@@ -4,9 +4,11 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/nobl9/govy/internal"
 	"github.com/nobl9/govy/internal/assert"
 
 	"github.com/nobl9/govy/pkg/govy"
+	"github.com/nobl9/govy/pkg/rules"
 )
 
 func TestValidator(t *testing.T) {
@@ -125,9 +127,70 @@ func TestValidatorInferName(t *testing.T) {
     - test`)
 }
 
+func TestValidatorValidateSlice(t *testing.T) {
+	t.Run("no errors", func(t *testing.T) {
+		v := govy.New(
+			govy.For(func(m mockValidatorStruct) string { return "test" }).
+				WithName("test").
+				Rules(govy.NewRule(func(v string) error { return nil })),
+		)
+		err := v.ValidateSlice([]mockValidatorStruct{{}})
+		assert.NoError(t, err)
+	})
+
+	t.Run("errors", func(t *testing.T) {
+		v := govy.New(
+			govy.For(func(m mockValidatorStruct) string { return m.Field }).
+				WithName("Field").
+				Required(),
+		).WithName("mock")
+		errs := mustValidatorErrors(t, v.ValidateSlice([]mockValidatorStruct{
+			{Field: "0"},
+			{},
+			{Field: "2"},
+			{},
+		}))
+		assert.Require(t, assert.Len(t, errs, 2))
+		assert.Require(t, assert.Len(t, errs[0].Errors, 1))
+		assert.Equal(t, govy.ValidatorErrors{
+			{
+				Name:       "mock",
+				SliceIndex: ptr(1),
+				Errors: govy.PropertyErrors{
+					&govy.PropertyError{
+						PropertyName: "Field",
+						Errors: []*govy.RuleError{{
+							Message: internal.RequiredErrorMessage,
+							Code:    rules.ErrorCodeRequired,
+						}},
+					},
+				},
+			},
+			{
+				Name:       "mock",
+				SliceIndex: ptr(3),
+				Errors: govy.PropertyErrors{
+					&govy.PropertyError{
+						PropertyName: "Field",
+						Errors: []*govy.RuleError{{
+							Message: internal.RequiredErrorMessage,
+							Code:    rules.ErrorCodeRequired,
+						}},
+					},
+				},
+			},
+		}, errs)
+	})
+}
+
 func mustValidatorError(t *testing.T, err error) *govy.ValidatorError {
 	t.Helper()
 	return mustErrorType[*govy.ValidatorError](t, err)
+}
+
+func mustValidatorErrors(t *testing.T, err error) govy.ValidatorErrors {
+	t.Helper()
+	return mustErrorType[govy.ValidatorErrors](t, err)
 }
 
 type mockValidatorStruct struct {
